refactor: 상품 목록 화면 분리
