+++ app/src/main/java/woowacourse/shopping/MainActivity.kt
@@ -1,19 +1,9 @@
 package woowacourse.shopping
 
-import ProductListAppBar
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import woowacourse.shopping.constants.MockData
 import woowacourse.shopping.ui.theme.AndroidshoppingTheme
 
 class MainActivity : ComponentActivity() {
@@ -24,27 +14,6 @@ class MainActivity : ComponentActivity() {
             AndroidshoppingTheme {
                 ProductListScreen()
             }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ProductListScreen() {
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        topBar = {
-            ProductListAppBar()
-        },
-    ) { innerPadding ->
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(innerPadding),
-        ) {
-            ProductList(
-                products = MockData.MOCK_PRODUCTS,
-            )
         }
     }
 }
